atualização final inventario

+++ CollectorQi/Models/ESCL028/ValidarReparosNotaFiscal.cs
@@ -5,8 +5,8 @@ namespace CollectorQi.Models.ESCL028
     public class ValidarReparosNotaFiscal
     {
         public string CodEstabel { get; set; } = "";
-        public int Digito { get; set; } = 0;
-        public int NumRR { get; set; } = 0;
+        public decimal Digito { get; set; } = 0;
+        public decimal NumRR { get; set; } = 0;
         public bool Conferido { get; set; } = false;
         public string CodFilial { get; set; } = "";
         public string Mensagem { get; set; } = "";
